update validation for rate value and others, use other means for testing without using the phpunit xml

+++ app/Interfaces/CurrencyHourlyConverterInterface.php
@@ -4,7 +4,7 @@ namespace App\Interfaces;
 
 interface CurrencyHourlyConverterInterface
 {
-    public function  convertHourlyRateBasedOnCurrency(?string $userCurrency, float $amount);
+    public function  convertHourlyRateBasedOnCurrency(?string $userCurrency, int $amount);
 
 
 }
